Update TfsClientProvider.cs
I ran into a case where the edit+rename flags where set on a file and the code crashed here. The above change fixed the problem. This was TFS 2008 and I think it happened because a folder was renamed and a file in the folder was deleted in the same change set.
One could possibly fix this more elegantly by executing the delete block before the rename block, but I fear this could have other consequences.